requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.10
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.11
django-filter==21.1
django-import-export==2.7.1
django-rest-framework==0.1.0
django-simpleui==2022.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
et-xmlfile==1.1.0
idna==3.3
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.9
psutil==5.9.0
pyasn1==0.4.8
pygame==2.0.1
PyJWT==2.3.0
PyMySQL==1.0.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
rsa==4.8
simplepro==3.3.5
simpleui==4.0.2
snowflake==0.0.3
sp==1.1.0
sqlparse==0.4.2
tablib==3.1.0
typing_extensions==4.0.1
urllib3==1.26.8
xlrd==2.0.1
xlwt==1.3.0
